requirements.txt
# 配置
PyYAML>=6.0

# 数据处理与数值
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0

# 步态/IMU 体轴转换（特征提取）
mobgap>=0.1.0

# 机器学习与训练
scikit-learn>=1.0.0
xgboost>=1.6.0
imbalanced-learn>=0.10.0
optuna>=3.0.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.12.0

# 模型可解释性（可选）
# shap>=0.43.0
